==> src/6-events-filtering.ipynb <==
# Events filtering
1. Create data with feasible agents;
2. Reproject events to make sure link_osm and link_id are consistent.

%load_ext autoreload
%autoreload 2
%cd D:\sysmo-data
%matplotlib inline

from tqdm.notebook import tqdm
import pandas as pd
import pickle
import os
os.environ['USE_PYGEOS'] = '0'
import geopandas as gpd

## 1. Load feasible agent IDs

df_id = pd.read_csv('dbs/results/combined_Sweden_all_unique_agents.csv')
feasible_agents = df_id.loc[df_id.agent_status == 'feasible', 'PId'].unique()
len(feasible_agents)

df_pop_s = pd.read_pickle("dbs/scenarios/syn_pop_s.pkl")
df_pop = pd.read_pickle('dbs/syn_pop_all.pkl')
df_pop = pd.merge(df_pop, df_id, on='PId', how='left')
df_pop.fillna('non_car_agent', inplace=True)
df_pop.drop(columns=['pot_car_driver'], inplace=True)
tqdm.pandas()
df_pop.loc[:, 'agent_status'] = df_pop.progress_apply(lambda row: 'stationary' if row['PId'] in df_pop_s.PId else row['agent_status'], axis=1)
df_pop.iloc[0]

df_pop.groupby('agent_status').size()

df_pop.to_parquet('dbs/results/syn_pop_all.parquet', index=False)

## 2. Filter plans

def process_plans(df=None):
    df_r_h = df.loc[(df['act_id'] == 0) & (df['src']=='input'), :]
    home_dict = dict(zip(df_r_h.PId, 
                         list(df_r_h[['POINT_X', 'POINT_Y']].itertuples(index=False, name=None))))
    df = df.loc[df['src'] == 'output', :]
    df.loc[:, 'POINT_X'] = df.apply(lambda row: home_dict[row['PId']][0] if row['act_id'] == 0 else row['POINT_X'],
                                        axis=1)
    df.loc[:, 'POINT_Y'] = df.apply(lambda row: home_dict[row['PId']][1] if row['act_id'] == 0 else row['POINT_Y'],
                                        axis=1)
    return df.drop(columns=['src'])

feasible_agents = df_pop.loc[df_pop.agent_status=='feasible', 'PId'].unique()
len(feasible_agents)

list_df = []
for region in ('palt20', 'palt21', 'palt22', 'palt23', 'palt24', 'palt25',
               'vast13', 'vast14', 'vast17',
                'samm', 'skane', 'sydost'):
    print(f'Process {region}...')
    df_r = pd.read_csv(f'dbs/output/plans_{region}.csv.gz', compression='gzip')
    df_r = process_plans(df_r)
    df_r = df_r.loc[df_r.PId.isin(feasible_agents), :]
    df_r.loc[:, 'region'] = region
    list_df.append(df_r)
df = pd.concat(list_df)

df.to_parquet('dbs/results/plans.parquet', index=False)

## 3. Filter events in line with the updated road network

gdf_network = gpd.read_file('dbs/network/network_o.shp')
gdf_network.head()

link_dict_rgs = dict()
for region in ('palt20', 'palt21', 'palt22', 'palt23', 'palt24', 'palt25',
               'vast13', 'vast14', 'vast17',
                'samm', 'skane', 'sydost'):
    d = gdf_network.loc[gdf_network.region == region, :]
    link_dict = dict(
        zip(d['link_osm'],
            list(d[['from_node', 'to_node']].itertuples(index=False, name=None)))
    )
    link_dict_rgs[region] = link_dict

link_dict = dict(
    zip(gdf_network['link_osm'],
        list(gdf_network[['from_node', 'to_node']].itertuples(index=False, name=None)))
)

def process_events(events=None, region=None):
    df = events.loc[events.person.isin(feasible_agents), :].copy()
    df.drop(columns=['link_id'], inplace=True)
    df.loc[:, 'from_node'] = df['link'].apply(lambda x: 
                                              link_dict[x][0] if x not in link_dict_rgs[region]
                                              else link_dict_rgs[region][x][0])
    df.loc[:, 'to_node'] = df['link'].apply(lambda x: 
                                            link_dict[x][1] if x not in link_dict_rgs[region] 
                                            else link_dict_rgs[region][x][1])
    df.loc[:, 'region'] = region
    return df

# Data location
for region in ('palt20', 'palt21', 'palt22', 'palt23', 'palt24', 'palt25',
               'vast13', 'vast14', 'vast17',
                'samm', 'skane', 'sydost'):
    print(f'Process {region}...')
    for batch in tqdm(range(0, 7), desc='Reading batches'):
        events_file = f'dbs/events/{region}_events_batch{batch}.csv.gz'
        df_events = pd.read_csv(events_file, compression='gzip')
        to_file = f'dbs/results/events/{region}_events_batch{batch}.parquet'
        df_proc = process_events(events=df_events, region=region)
        df_proc.to_parquet(to_file, index=False)
