==> src/7-data-product-description.ipynb <==
# Describe data products
1. Synthetic population attributes
2. Activity plans
3. Car agents' driving trajectories
4. Road network

%load_ext autoreload
%autoreload 2
%cd D:\sysmo-data
%matplotlib inline

from tqdm.notebook import tqdm
import pandas as pd
import pickle
import os
os.environ['USE_PYGEOS'] = '0'
import geopandas as gpd

## 1. Synthetic agents

df_pop = pd.read_parquet('dbs/results/syn_pop_all.parquet')
print(df_pop.head())

df_pop.groupby('agent_status').size()

## 2. Activity plans

df_plans = pd.read_parquet('dbs/results/plans.parquet')
print(df_plans.head())

## 3. Car agents' driving trajectories
All files are under events folder where each region's data is divided into 7 batches by individual agents.

df_traj = pd.read_parquet('dbs/results/events/palt20_events_batch0.parquet')
print(df_traj.head())

## 4. Road network 
With geometry corresponding to 'link' column in the trajectories.

gdf_network = gpd.read_file('dbs/results/network_o/network_o.shp')
print(gdf_network.head())
